+++ Changelog.rst
@@ -24,6 +24,7 @@ Breaks in compatibility
 * The ``counter`` parameter of a CTR mode cipher must be generated via
   ``Crypto.Util.Counter``. It cannot be a generic callable anymore.
 * Removed the ``Crypto.Random.Fortuna`` package (due to lack of test vectors).
+* Removed the ``Crypto.Hash.new`` function.
 
 3.0
 +++